Add files via upload
This is adding a function that adds the existing CSV file to a Dataframe, then looks it up to figure out if the Seek record has already been queried. It then appends any new entries to the end of the CSV file. 
If the CSV file doesn't exist, it will query everything and create the file with headings. 

The lookup was relatively straight forward - joining it all together proved to be challenging as I only want to load the CSV file to the Dataframe once. Not for each query as I did for testing. The issue here was I had to use global variables, and I struggled to get my head around them with Jupyter Notebook (as sometimes you had to clear output for variables to update). Still not 100% confident on some theory, and quite a few bugs to fix. But I got there in the end. 

20% of time was creating the lookup, the remaining 80% was getting global variables to work!